Implement proxy event ordering and fix generated method names

+++ src/main/java/com/kmecpp/osmium/api/event/Order.java
@@ -38,12 +38,12 @@ public enum Order implements Abstraction {
 			LAST.source = org.spongepowered.api.event.Order.LAST;
 			POST.source = org.spongepowered.api.event.Order.POST;
 		} else if (Platform.isProxy()) {
-			FIRST.source = net.md_5.bungee.event.EventPriority.LOWEST - 1;
-			EARLY.source = net.md_5.bungee.event.EventPriority.LOW;
-			DEFAULT.source = net.md_5.bungee.event.EventPriority.NORMAL;
-			LATE.source = net.md_5.bungee.event.EventPriority.HIGH;
-			LAST.source = net.md_5.bungee.event.EventPriority.HIGHEST;
-			POST.source = net.md_5.bungee.event.EventPriority.HIGHEST + 1;
+			FIRST.source = (byte) (net.md_5.bungee.event.EventPriority.LOWEST - 1);
+			EARLY.source = (byte) (net.md_5.bungee.event.EventPriority.LOW);
+			DEFAULT.source = (byte) (net.md_5.bungee.event.EventPriority.NORMAL);
+			LATE.source = (byte) (net.md_5.bungee.event.EventPriority.HIGH);
+			LAST.source = (byte) (net.md_5.bungee.event.EventPriority.HIGHEST);
+			POST.source = (byte) (net.md_5.bungee.event.EventPriority.HIGHEST + 1);
 		}
 	}
